pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "chemstack"
version = "0.1.0"
description = "ChemStack monorepo for ORCA, xTB, CREST, and workflow automation"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
    { name = "Daehyup Sohn" },
]
keywords = [
    "orca",
    "dft",
    "computational-chemistry",
    "workflow-automation",
    "cli",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Chemistry",
]
dependencies = [
    "PyYAML>=6.0.1",
]

[project.urls]
Homepage = "https://github.com/dhsohn/chemstack"
Repository = "https://github.com/dhsohn/chemstack"
Issues = "https://github.com/dhsohn/chemstack/issues"

[project.optional-dependencies]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.5",
    "types-PyYAML>=6.0.12",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["chemstack*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.mypy]
python_version = "3.10"
files = ["src/chemstack", "tests"]
pretty = true
show_error_codes = true
warn_unused_configs = true

[tool.coverage.run]
source = ["src/chemstack"]

[tool.coverage.report]
fail_under = 80
show_missing = true
